removed jekyll-tagging plugin
I figured the jekyll-tagging plugin didn't as intended on gh-pages, and
according to to the page below, I was right.

His method works, but .md files need to be
made for each tag. The fellow offers an autogenerator, but it doesn't
work.

http://longqian.me/2017/02/09/github-jekyll-tag/

@up-for-grabs uses some kind of auto-generating script, as far as
I could tell, but I haven't studied it thoroughly yet.